Added support for eggs, raids, quests, invasion, ...
